remove arrow/constraint api. The usage of PConstant is unlawful
these ideas live on in https://github.com/mlabs-haskell/apropos-tx/pull/33
