+++ apropos-tx.cabal
@@ -62,15 +62,8 @@ common lang
 
 library
   import:          lang
-  exposed-modules: Apropos.HasResourceBounds
-                 , Apropos.Script
-                 , Apropos.Script.Model
-                 , Apropos.Script.Iso.Arrow
-                 , Apropos.Script.Iso.Arrow.Runner
-                 , Apropos.Script.Iso.Constraint
-                 , Apropos.Script.Iso.Constraint.Runner
+  exposed-modules: Apropos.Script
                  , Apropos.Tx
-                 , Apropos.Tx.Constraint.Policy
   hs-source-dirs:  src
 
 test-suite examples
@@ -81,9 +74,6 @@ test-suite examples
   ghc-options: -threaded -rtsopts -with-rtsopts=-N
   other-modules:
     Spec.Int
-    Spec.IntArrow
-    Spec.IntArrowConstraint
-    Spec.IntConstraint
     Spec.IntPermutationGen
     Spec.IntPair
     Spec.Plutarch.MagicNumber
