+++ examples/Main.hs
@@ -1,9 +1,6 @@
 module Main (main) where
 
 import Spec.Int
-import Spec.IntArrow
-import Spec.IntArrowConstraint
-import Spec.IntConstraint
 import Spec.IntPair
 import Spec.IntPermutationGen
 import Spec.Plutarch.CostModel
@@ -42,17 +39,5 @@ tests =
         [ magicNumberPropGenTests
         , addCostPropGenTests
         , addCostModelPlutarchTests
-        ]
-    , testGroup
-        "Arrow test"
-        [ intArrowPlutarchTests
-        ]
-    , testGroup
-        "Constraint Test"
-        [ intConstraintPlutarchTests
-        ]
-    , testGroup
-        "Arrow and Constraint composition Test"
-        [ intArrowConstraintPlutarchTests
         ]
     ]
